requirements.txt
anthropic
rich
python-pptx
requests
python-dotenv
